fix: Resolve all compilation errors for GitHub Actions build
Fixed 28 compilation errors that were causing GitHub Actions build failures:

Type Mismatch Fixes (Int → Double):
- AnimeCard.kt: Updated getScoreColor() to accept Double? parameter
- AnimeRepository.kt: Convert fallback score Int to Double via toDouble()

Unresolved Reference Fixes:
- NavigationBar.kt: Added missing FontWeight import
- DetailScreen.kt: Added Warning/Refresh icon imports, fixed Surface onClick
- SearchScreen.kt: Added Brush import, fixed padding params, use Outlined.Close
- AppUpdater.kt: Removed BuildConfig dependency, use PackageManager for version
- UpdateDialog.kt: Replaced non-existent icons with valid alternatives:
  * Download → Icons.Default.FileDownload / Icons.Default.Download
  * Update → Icons.Default.SystemUpdate
  * OpenInNew → Icons.Default.OpenInNew / Icons.Outlined.OpenInNew
  * CheckCircle → Icons.Default.CheckCircle
  * InstallMobile → Icons.Default.Download
  * Added proper coroutineScope and imports for launch/delay/Intent/Uri

All changes ensure compatibility with Material Icons library and proper
Kotlin coroutines usage within Compose context.

+++ app/src/main/java/com/anitrack/app/data/repository/AnimeRepository.kt
@@ -302,7 +302,7 @@ class AnimeRepository private constructor(
                 medium = "https://s4.anilist.co/file/anilistcdn/media/anime/cover/small/${id}.jpg"
             ),
             episodes = episodes,
-            averageScore = score,
+            averageScore = score.toDouble(),
             genres = genres,
             status = status,
             seasonYear = 2024
