+++ app/src/main/java/com/anitrack/app/ui/favorites/FavoritesScreen.kt
@@ -384,12 +384,12 @@ private fun FavoritesHeader() {
     }
 }
 
-private fun getScoreColor(score: Int?): Color {
+private fun getScoreColor(score: Double?): Color {
     return when {
         score == null -> RatingAverage
-        score >= 80 -> RatingExcellent
-        score >= 60 -> RatingGood
-        score >= 40 -> RatingAverage
+        score >= 80.0 -> RatingExcellent
+        score >= 60.0 -> RatingGood
+        score >= 40.0 -> RatingAverage
         else -> RatingPoor
     }
 }
